cs compiler: map IntType to int

+++ tests/machine/x/cs/fun_expr_in_let.cs
@@ -1,13 +1,10 @@
 using System;
 
-class Program
-{
-    static void Main()
-    {
-        Func<long, long> square = new Func<long, long>((long x) =>
-        {
-            return (x * x);
-        });
+class Program {
+    static void Main() {
+        Func<int, int> square = new Func<long, long>((long x) => {
+    return (x * x);
+});
         Console.WriteLine(square(6));
     }
 }
