+++ tests/machine/x/cs/if_else.cs
@@ -2,7 +2,7 @@ using System;
 
 class Program {
     static void Main() {
-        long x = 5;
+        int x = 5;
         if (x > 3) {
             Console.WriteLine("big");
         } else {
@@ -10,4 +10,3 @@ class Program {
         }
     }
 }
-
