+++ tests/machine/x/cs/let_and_print.cs
@@ -1,10 +1,8 @@
 using System;
 
-class Program
-{
-    static void Main()
-    {
-        long a = 10;
+class Program {
+    static void Main() {
+        int a = 10;
         long b = 20;
         Console.WriteLine((a + b));
     }
