+++ tests/machine/x/cs/map_assign.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 class Program {
     static void Main() {
-        Dictionary<string, long> scores = new Dictionary<string, long> { { "alice", 1 } };
+        Dictionary<string, int> scores = new Dictionary<string, int> { { "alice", 1 } };
         scores["bob"] = 2;
         Console.WriteLine(scores["bob"]);
     }
