+++ tests/machine/x/cs/map_index.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 class Program {
     static void Main() {
-        Dictionary<string, long> m = new Dictionary<string, long> { { "a", 1 }, { "b", 2 } };
+        Dictionary<string, int> m = new Dictionary<string, int> { { "a", 1 }, { "b", 2 } };
         Console.WriteLine(m["b"]);
     }
 }
