+++ tests/machine/x/cs/map_nested_assign.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 
 class Program {
     static void Main() {
-        Dictionary<string, Dictionary<string, long>> data = new Dictionary<string, Dictionary<string, long>> { { "outer", new Dictionary<string, long> { { "inner", 1 } } } };
+        Dictionary<string, Dictionary<string, int>> data = new Dictionary<string, Dictionary<string, int>> { { "outer", new Dictionary<string, int> { { "inner", 1 } } } };
         data["outer"]["inner"] = 2;
         Console.WriteLine(data["outer"]["inner"]);
     }
